Update plugin.xml
Having android-minSdkVersion at 15 was causing a build problem in projects with the latest version of cordova-plugin-fcm-with-dependecy-updated.  Updating it to 16 fixes this build problem.  Incidentally, it also fixes a visual issue where the app doesn't respect the spacing of the top notches on certain ios devices.  Updating to 16 also fixes this issue, though I can't explain why.